+++ tests/swagger_generation/registries/hidden_api.py
@@ -222,23 +222,31 @@ EXPECTED_SWAGGER_V2 = {
     },
     "definitions": {
         "Foo": {
+            "additionalProperties": False,
             "type": "object",
             "title": "Foo",
             "properties": {"uid": {"type": "string"}, "name": {"type": "string"}},
         },
         "FooUpdateSchema": {
+            "additionalProperties": False,
             "type": "object",
             "title": "FooUpdateSchema",
             "properties": {"name": {"type": "string"}},
         },
         "NestedFoosSchema": {
+            "additionalProperties": False,
             "type": "object",
             "title": "NestedFoosSchema",
             "properties": {
-                "data": {"type": "array", "items": {"$ref": "#/definitions/Foo"}}
+                "data": {
+                    "additionalProperties": False,
+                    "type": "array",
+                    "items": {"$ref": "#/definitions/Foo"},
+                }
             },
         },
         "Error": {
+            "additionalProperties": False,
             "type": "object",
             "title": "Error",
             "properties": {"message": {"type": "string"}, "errors": {"type": "object"}},
@@ -255,26 +263,31 @@ SWAGGER_V3_WITHOUT_HIDDEN = expected_swagger = {
     "components": {
         "schemas": {
             "Foo": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "Foo",
                 "properties": {"uid": {"type": "string"}, "name": {"type": "string"}},
             },
             "FooUpdateSchema": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "FooUpdateSchema",
                 "properties": {"name": {"type": "string"}},
             },
             "NestedFoosSchema": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "NestedFoosSchema",
                 "properties": {
                     "data": {
+                        "additionalProperties": False,
                         "type": "array",
                         "items": {"$ref": "#/components/schemas/Foo"},
                     }
                 },
             },
             "Error": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "Error",
                 "properties": {
@@ -428,26 +441,31 @@ SWAGGER_V3_WITH_HIDDEN = expected_swagger = {
     "components": {
         "schemas": {
             "Foo": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "Foo",
                 "properties": {"uid": {"type": "string"}, "name": {"type": "string"}},
             },
             "FooUpdateSchema": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "FooUpdateSchema",
                 "properties": {"name": {"type": "string"}},
             },
             "NestedFoosSchema": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "NestedFoosSchema",
                 "properties": {
                     "data": {
+                        "additionalProperties": False,
                         "type": "array",
                         "items": {"$ref": "#/components/schemas/Foo"},
                     }
                 },
             },
             "Error": {
+                "additionalProperties": False,
                 "type": "object",
                 "title": "Error",
                 "properties": {