+++ tests/test_validation.py
@@ -25,14 +25,8 @@ from flask_rebar import rebar
 from flask_rebar.utils.request_utils import normalize_schema
 from flask_rebar.validation import RequireOnDumpMixin
 from flask_rebar.validation import CommaSeparatedList
-from flask_rebar.validation import DisallowExtraFieldsMixin
 from flask_rebar.validation import QueryParamList
 from tests.test_rebar import create_rebar_app
-
-
-class DisallowExtraFieldsSchema(Schema, DisallowExtraFieldsMixin):
-    a = fields.String()
-    b = fields.String(load_from="c")
 
 
 class NoRequireOnDumpMixinSchema(Schema):